Minor fixing

+++ builder/shell/shell.go
@@ -4,7 +4,7 @@ type InstanceType struct {
 	ServerName           string
 	Hypervisor           string
 	Cpus                 int
-	Memory               int
+	Memory               int64
     DiskSize             int
 	OsName               string
 	Network 			 string
